Track NodeInfo packet count per node
